requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Faker==37.0.2
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
